Skip markdown comments
